project/requirements.txt
streamlit==1.36.0
Pillow==10.4.0
pandas==2.2.2
requests==2.32.3
boto3==1.34.131
python-dotenv==1.0.1
streamlit-image-zoom==0.2.0